Moved user to own file and simplified getting them
